requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12.0,<4.0
psycopg2-binary>=2.8.6
django-cors-headers>=3.7.0
Pillow>=8.0.0
django-crispy-forms>=1.11.0
